requirements.txt
uvicorn==0.21.1
sentence-transformers==2.2.2
faiss-cpu==1.7.3
fastapi==0.95.0
onnxruntime==1.14.1
openpyxl==3.1.2
python-dotenv==1.0.0
pandas==1.5.3
gunicorn==20.1.0
SQLAlchemy==2.0.9
psycopg2-binary==2.9.6